Move all Volume related helper methods into VolumeUtils.
